Move ORMContentTypeEntriesProxy to its own file.
